Generate seed data.
Someone new checking out the repo can get started with sending GET
request and be able to play with the url to see jsonapi in action. This
speeds up the ability to try out the features without forcing the person
to commit to making up multiple request to create a meaningful dataset.